backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic>=2.0
pydantic-settings
python-dotenv
google-genai>=0.2.0
psycopg2-binary
redis
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy
python-multipart
alembic
email-validator